Add files via upload
Chef is participating in a race of Y kilometers. However, chef gets tired and needs to take rest every X kilometers. We need to determine the number of times Chef stops to take rest before reaching the finish line.

We will take 2 inputs - X and Y.
Initialize count =0. this will count the number of stops.
And then we will run a loop which will start from X and end just before reaching Y with every increment of X in it. In the loop count will be incremented by 1.
After the loop, we will display the count.