Add files via upload
Chef and Chefina are ordering pizza.
Each pizza has N slices.
what is the minimum number of pizza they must order so that both can eat equal number of slices?

If input - 6 then output will be 1.
If input - 3 then output will be 2.  

Solution
For that we will have an if condition where modulo operation will be applied on input. N%2==0 then we will print 1.
else print 2.